+++ django_messages_drf/tests/utils.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+
+
+class SerializerTest(serializers.Serializer):
+    """
+    Test Serializer
+    """
+    name = serializers.CharField(required=False)